+++ src/Filesystem.php
@@ -39,6 +39,19 @@ class Filesystem extends BaseFilesystem
     }
 
     /**
+     * Concatenates various paths together, adding the right slash term
+     * @param string $paths Various paths to be concatenated
+     * @return string The path concatenated
+     * @since 1.4.5
+     */
+    public function concatenate(string ...$paths): string
+    {
+        $end = array_pop($paths);
+
+        return implode('', array_map([$this, 'addSlashTerm'], $paths)) . $end;
+    }
+
+    /**
      * Creates a file. Alias for `mkdir()` and `file_put_contents()`.
      *
      * It also recursively creates the directory where the file will be created.
@@ -201,7 +214,18 @@ class Filesystem extends BaseFilesystem
             return $endPath;
         }
 
-        return $this->addSlashTerm($startPath) . $endPath;
+        return $this->concatenate($startPath, $endPath);
+    }
+
+    /**
+     * Normalizes the path, applying the right slash term
+     * @param string $path Path you want normalized
+     * @return string Normalized path
+     * @since 1.4.5
+     */
+    public function normalizePath(string $path): string
+    {
+        return str_replace(['/', '\\'], DS, $path);
     }
 
     /**
@@ -281,10 +305,10 @@ class Filesystem extends BaseFilesystem
     {
         $root = $this->getRoot();
         if ($this->isAbsolutePath($path) && string_starts_with($path, $root)) {
-            $path = $this->makePathRelative($path, $root);
+            $path = $this->normalizePath($this->makePathRelative($path, $root));
         }
 
-        return rtrim($path, '/');
+        return rtrim($path, DS);
     }
 
     /**