+++ tests/TestCase/FilesystemTest.php
@@ -53,6 +53,18 @@ class FilesystemTest extends TestCase
     }
 
     /**
+     * Test for `concatenate()` method
+     * @test
+     */
+    public function testConcatenate()
+    {
+        $this->assertSame('dir', $this->Filesystem->concatenate('dir'));
+        $this->assertSame('dir' . DS . 'subdir', $this->Filesystem->concatenate('dir', 'subdir'));
+        $this->assertSame('dir' . DS . 'subdir', $this->Filesystem->concatenate('dir' . DS, 'subdir'));
+        $this->assertSame('dir' . DS . 'subdir' . DS . 'subsubdir', $this->Filesystem->concatenate('dir', 'subdir', 'subsubdir'));
+    }
+
+    /**
      * Test for `createFile()` method
      * @test
      */
@@ -202,20 +214,6 @@ class FilesystemTest extends TestCase
     }
 
     /**
-     * Test for `makePathAbsolute()` method
-     * @test
-     */
-    public function testMakePathAbsolute()
-    {
-        $this->assertSame(TMP . 'dir', $this->Filesystem->makePathAbsolute(TMP . 'dir', TMP));
-        $this->assertSame(TMP . 'dir', $this->Filesystem->makePathAbsolute('dir', TMP));
-
-        $this->expectException(InvalidArgumentException::class);
-        $this->expectExceptionMessage('The start path `relativePath` is not absolute');
-        $this->Filesystem->makePathAbsolute('dir', 'relativePath');
-    }
-
-    /**
      * Test for `isSlashTerm()` method
      * @test
      */
@@ -262,6 +260,34 @@ class FilesystemTest extends TestCase
     }
 
     /**
+     * Test for `makePathAbsolute()` method
+     * @test
+     */
+    public function testMakePathAbsolute()
+    {
+        $this->assertSame(TMP . 'dir', $this->Filesystem->makePathAbsolute(TMP . 'dir', TMP));
+        $this->assertSame(TMP . 'dir', $this->Filesystem->makePathAbsolute('dir', TMP));
+
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('The start path `relativePath` is not absolute');
+        $this->Filesystem->makePathAbsolute('dir', 'relativePath');
+    }
+
+    /**
+     * Test for `normalizePath()` method
+     * @test
+     */
+    public function testNormalizePath()
+    {
+        foreach ([
+            'path/to/normalize',
+            'path\\to\\normalize',
+        ] as $path) {
+            $this->assertSame('path' . DS . 'to' . DS . 'normalize', $this->Filesystem->normalizePath($path));
+        }
+    }
+
+    /**
      * Test for `rmdirRecursive()` method
      * @test
      */
@@ -283,11 +309,11 @@ class FilesystemTest extends TestCase
      */
     public function testRtr()
     {
-        $this->assertSame('my/folder', $this->Filesystem->rtr(ROOT . 'my' . DS . 'folder'));
+        $this->assertSame('my' . DS . 'folder', $this->Filesystem->rtr(ROOT . 'my' . DS . 'folder'));
 
         //Resets the ROOT value, removing the final slash
         putenv('ROOT=' . rtrim(ROOT, DS));
-        $this->assertSame('my/folder', $this->Filesystem->rtr(ROOT . 'my' . DS . 'folder'));
+        $this->assertSame('my' . DS . 'folder', $this->Filesystem->rtr(ROOT . 'my' . DS . 'folder'));
     }
 
     /**