+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 1.x branch
 ## 1.4 branch
 ### 1.4.5
-* added `Filesystem::makePathAbsolute()` method.
+* added `concatenate()`, `makePathAbsolute()` and `normalizePath()` methods for
+    `Filesystem`.
 
 ### 1.4.4
 * added `Filesystem` class, all filesystem global function are now deprecated;